Renaming regression tests to integration
