+++ config.properties
@@ -4,7 +4,7 @@ maxLoc=10000
 minLoc=100
 downloadDir=database
 benchmarkDir=benchmarks
-debug=true
+debug=false
 debugLevel=2
 type=R
 minExpr=1
