+++ src/java/transform/Main.java
@@ -55,10 +55,10 @@ public class Main {
 	private static boolean debug = false;
 	private static boolean transformAll = false;
 
-//	 public static String source = "suitablePrgms";
-//	 public static String dest = "benchmarks";
-	 public static String source = "test/transformer/regression";
-	 public static String dest = "testOutput";
+	 public static String source = "suitablePrgms";
+	 public static String dest = "benchmarks";
+//	 public static String source = "test/transformer/integration";
+//	 public static String dest = "testOutput";
 
 	public static void main(String[] args) throws IOException {
 		File tmpDir = Files.createTempDirectory("paclab-transform").toFile();
@@ -209,7 +209,7 @@ public class Main {
 			}
 
 			// Create the YAML if targeting SVCOMP
-			if(target.equals("SVCOMP")) {
+			if(target.equals("SVCOMP") && success) {
 				createSVCompYmlFile(file);
 			}
 		});
